+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/Gold.java
@@ -9,28 +9,24 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.*;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.SwordItem;
+import net.minecraft.world.item.TieredItem;
+import net.minecraft.world.item.Tiers;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.common.TierSortingRegistry;
 import net.minecraftforge.event.entity.living.LootingLevelEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 
-@Label(name = "Gold", description = "Changes Gold tools to have an innate Fortune/Looting I and changes the harvest level to be like stone tools")
+@Label(name = "Gold", description = "Changes Gold tools to have an innate Fortune/Looting I")
 @LoadFeature(module = Modules.Ids.MINING)
 public class Gold extends Feature {
 
 	public static final String LUCKY_GOLD_TOOLTIP = InsaneSurvivalOverhaul.MOD_ID + ".innate_luck";
-
-	@Config
-	@Label(name = "Harvest Level", description = "Set the harvest level of gold tools. Vanilla is minecraft:gold (same as minecraft:wood), the there's stone, iron, diamond, netherite. Please note that an invalid resource location or harvest level here might crash the game.")
-	public static String harvestLevel = "minecraft:stone";
-	private static ResourceLocation _harvestLevel;
 
 	@Config(min = 0, max = 255)
 	@Label(name = "Looting Level", description = "Set the innate looting level of gold tools.")
@@ -46,7 +42,6 @@ public class Gold extends Feature {
 	@Override
 	public void readConfig(ModConfigEvent event) {
 		super.readConfig(event);
-		_harvestLevel = ResourceLocation.tryParse(harvestLevel);
 	}
 
 	@SubscribeEvent
@@ -57,7 +52,7 @@ public class Gold extends Feature {
 			return;
 
 		ItemStack stack = livingEntity.getMainHandItem();
-		if (stack.getItem() instanceof SwordItem swordItem && swordItem.getTier() == Tiers.GOLD) {
+		if (stack.getItem() instanceof SwordItem swordItem && swordItem.builtInRegistryHolder().is()) {
 			if (event.getLootingLevel() < lootingLevel)
 				event.setLootingLevel(lootingLevel);
 		}
@@ -71,15 +66,6 @@ public class Gold extends Feature {
 			return prev;
 
 		return fortuneLevel;
-	}
-
-	public static Tier getEffectiveTier(Tier originalTier) {
-		if(!isEnabled(Gold.class)
-				|| originalTier != Tiers.GOLD
-				|| TierSortingRegistry.getName(Tiers.GOLD) == _harvestLevel)
-			return originalTier;
-
-		return TierSortingRegistry.byName(_harvestLevel);
 	}
 
 	@OnlyIn(Dist.CLIENT)